backend/requirements.txt
beautifulsoup4==4.12.3
elasticsearch==8.16.0
feedparser==6.0.11
Flask==3.1.0
Flask_Cors==5.0.0
numpy==2.0.2
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
yfinance==0.2.50
apify-client==1.5.0
